% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/punycode.R
\name{is_idna}
\alias{is_idna}
\title{Test if domain is IDNA encoded}
\usage{
is_idna(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
logical vector (\code{TRUE} if IDNA test passed)
}
\description{
Given a vector of domains, test if the domain is IDNA encdoed
}
\examples{
is_idna("новый-год.com", "rudis.net",
        "xn----peurf0asz4dzaln0qm161er8pd.biz")
}

