% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/punycode.R
\name{is_punycode}
\alias{is_punycode}
\title{Test if domain is Punycode}
\usage{
is_punycode(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
logical vector (\code{TRUE} if Punycode test passed)
}
\description{
Given a vector of domains, test if the domain is Punycode encdoed
}
\examples{
is_punycode("новый-год.com", "rudis.net",
            "xn----peurf0asz4dzaln0qm161er8pd.biz")
}

