% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{puny_decode}
\alias{puny_decode}
\title{Decode IDNA domains to punycode}
\usage{
puny_decode(domains)
}
\arguments{
\item{domains}{character vector of IDNA domains}
}
\value{
a character vector of punycode domains
}
\description{
Converts IDNA domains to punycode
}
\examples{
\dontrun{
}
}
\seealso{
\url{http://www.faqs.org/rfcs/rfc3492.html}
}

