% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{puny_encode}
\alias{puny_encode}
\title{Encode punycode domains to IDNA}
\usage{
puny_encode(domains)
}
\arguments{
\item{domains}{character vector of IDNA domains}
}
\value{
a character vector of IDNA domains
}
\description{
Converts punycode domains to IDNA
}
\examples{
puny_encode("xn------qpeiobbci9acacaca2c8a6ie7b9agmy.net")
}
\seealso{
\url{http://www.faqs.org/rfcs/rfc3492.html}
}

