% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/punycode-package.R
\docType{package}
\name{punycode}
\alias{punycode}
\alias{punycode-package}
\title{A package to handle punycode domains (encoding / decoding)}
\description{
It uses the GNU libidn \url{http://www.gnu.org/software/libidn/} library and
provides functions to encode & decode international domain names.
}
\author{
Bob Rudis (@hrbrmstr)
}

