% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/punycode.R
\name{is_idna}
\alias{is_idna}
\title{Test if domain is IDNA encoded}
\usage{
is_idna(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
logical vector (\code{TRUE} if IDNA test passed)
}
\description{
Given a vector of domains, test if the domain is IDNA encdoed
}
\examples{
\dontrun{
# See the vignette since character set problems prevent clean latex
# function example builds
vignette("introduction_to_punycode")
}
# see the vignette
}

