% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{puny_encode}
\alias{puny_encode}
\title{Encode Punycode domains to IDNA}
\usage{
puny_encode(domains)
}
\arguments{
\item{domains}{character vector of IDNA domains}
}
\value{
a character vector of IDNA domains
}
\description{
Converts Punycode domains to IDNA.
}
\note{
If there are any issues during the conversion, \code{"Invalid"} will
      be the value returned.
}
\examples{
# see the vignette
}
\seealso{
\url{http://www.faqs.org/rfcs/rfc3492.html} and
    \url{http://www.faqs.org/rfcs/rfc3490.html} and
    \url{http://www.gnu.org/software/libidn/}
}

