% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{puny_tld_check}
\alias{puny_tld_check}
\title{Check domains for validity}
\usage{
puny_tld_check(domains)
}
\arguments{
\item{domains}{character vector of domains}
}
\value{
a logical vector
}
\description{
Check international domain names for charset validity
}
\examples{
# see the vignette
}
\seealso{
\url{http://www.faqs.org/rfcs/rfc3492.html} and
    \url{http://www.faqs.org/rfcs/rfc3490.html} and
    \url{http://www.gnu.org/software/libidn/}
}

