% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/punycode-package.R
\docType{package}
\name{punycode}
\alias{punycode}
\alias{punycode-package}
\title{A package to handle punycode domains (encoding /)}
\description{
A package to handle punycode domains (encoding /)
}
\author{
Bob Rudis (@hrbrmstr)
}

